MGNL-14935 Proberten te reproduceren. Wat clean up en simplificataties. JMX call uitgebreid zodat je er makkelijker ééntje kunt testen.

+++ vpro-shared-util/src/main/java/nl/vpro/util/FileCachingInputStream.java
@@ -180,12 +180,14 @@ public class FileCachingInputStream extends InputStream {
         }
         if (file != null) {
             file.close();
-            if (Files.deleteIfExists(tempFile)) {
-                log.debug("Deleted {}", tempFile);
-            } else {
-                log.warn("Could not delete {}", tempFile);
+            if (tempFile != null) {
+                if (Files.deleteIfExists(tempFile)) {
+                    log.debug("Deleted {}", tempFile);
+                } else {
+                    //   openOptions.add(StandardOpenOption.DELETE_ON_CLOSE); would have arranged that!
+                    log.debug("Could not delete because didn't exists any more {}", tempFile);
+                }
             }
-
         }
     }
